refactor: share the drive file processor, large-file pipeline and change classifier
Closes #306
