+++ packages/drive/src/index.ts
@@ -1,5 +1,27 @@
+/**
+ * The drive behaviour OneDrive and SharePoint share, taking each provider's values as arguments.
+ *
+ * The modules that still exist twice in the provider packages fall into two groups (#306). The
+ * ones with real differences carry a comment saying so and what the difference is:
+ * `backup.service.ts`, `restore.service.ts`, `status.service.ts`, `backup-builders.ts` and
+ * `restore-folder-path.ts`. Everything else that still pairs up by name is a binder: a class that
+ * satisfies a per-provider port, a descriptor that names the manifest lookup, or a module that
+ * supplies the key layout and a chunk fetcher to the shared code below. Those exist precisely to
+ * hold the provider's values, so deduplicating them would mean giving up the separate ports, and
+ * a near-zero diff there is the intended result rather than duplication left behind.
+ */
 export * from '@/drive-ports';
 export { download_with_retry, type DownloadRetryOptions } from '@/backup/download-retry';
+export { classify_drive_change } from '@/backup/change-classifier';
+export { process_drive_backup_file, type FileProcessResult } from '@/backup/file-processor';
+export {
+  process_large_drive_file,
+  cleanup_stale_drive_staging,
+  type DriveDownloadUrlResolver,
+  type DriveLargeFileDeps,
+  type LargeFileResult,
+} from '@/backup/large-file-pipeline';
+export { format_bytes } from '@/shared/format-bytes';
 export { LARGE_FILE_THRESHOLD } from '@/backup/large-file-threshold';
 export {
   stream_whole_file_in_chunks,
