+++ packages/onedrive/src/services/restore/restore-folder-path.ts
@@ -1,3 +1,6 @@
+// Deliberately not shared with the SharePoint copy (#306): 28 differing lines of 106. The memo
+// keys differ, path here against `drive_id:path` there, which is a behavioural difference
+// rather than a naming one and is tracked separately.
 import type { OneDriveConnector } from '@wisecom/atlas-types';
 import { logger } from '@wisecom/atlas-core/utils/logger';
